+++ wc-invoice.php
@@ -4,7 +4,7 @@
  * Plugin Name: WC Invoice
  * Plugin URI: https://github.com/ahmadreza-log/wc-invoice
  * Description: Professional invoice generator plugin for WooCommerce with PDF and HTML invoice generation capabilities
- * Version: 0.0.11
+ * Version: 0.0.12
  * Author: Ahmadreza Ebrahimi
  * Author URI: https://ahmadreza.me
  * License: GPL v2 or later
@@ -26,12 +26,13 @@ use WC_Invoice\Ajax;
 use WC_Invoice\Addon_Manager;
 use WC_Invoice\Database;
 use WC_Invoice\Enqueue;
+use WC_Invoice\Frontend;
 use WC_Invoice\Notices;
 use WC_Invoice\Woocommerce;
 use WC_Invoice\Generator;
 
 // Define plugin constants
-define('WC_INVOICE_VERSION', '0.0.11');
+define('WC_INVOICE_VERSION', '0.0.12');
 define('WC_INVOICE_FILE', __FILE__);
 define('WC_INVOICE_DIR', plugin_dir_path(WC_INVOICE_FILE));
 define('WC_INVOICE_URL', plugin_dir_url(WC_INVOICE_FILE));
@@ -148,6 +149,7 @@ class WC_Invoice
 
         Enqueue::instance();
         Ajax::instance();
+        Frontend::instance();
     }
 
     /**